requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
azure-mgmt-core
azure-mgmt-costmanagement==4.0.0b1
azure-storage-blob==12.15.0
azure-storage-file-datalake==12.10.1
urllib3==1.26.14
urlrequest==1.0.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
idna==3.4
isodate==0.6.1
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
pandas
portalocker==2.7.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.5.0
